Simplified github actions build #2830
