+++ github-actions/scan/src/action-helper.ts
