+++ github-actions/scan/src/cli-helper.ts
