+++ github-actions/scan/src/fs-helper.ts
