+++ github-actions/scan/src/log-helper.ts
