+++ github-actions/scan/src/report-formats.ts
