--- github-actions/shared/.prettierrc.json
@@ -1,6 +1,0 @@
-{
-  "printWidth": 150,
-  "semi": true,
-  "singleQuote": true,
-  "tabWidth": 4
-}
