--- github-actions/shared/jest.config.js
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: MIT
-
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-};
