--- github-actions/shared/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2018",
-    "moduleResolution": "node",
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "outDir": "dist"
-  }
-}
